requirements.txt
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
Django==2.0
djangorestframework==3.7.3
httpie==0.9.9
idna==2.6
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
urllib3==1.22
